Authorization Complete

+++ app/Http/Controllers/backend/RoleController.php
@@ -7,6 +7,7 @@ use App\Models\Module;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('app.roles.index');
         $roles = Role::all();
         return view('backend.roles.index',compact('roles'));
     }
@@ -28,6 +30,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('app.roles.create');
         $modules = Module::all();
         return view('backend.roles.form',compact('modules'));
     }
@@ -40,6 +43,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('app.roles.create');
         $this->validate($request,[
             'name'  =>  'required|unique:roles',
             'permissions'   =>  'required|array',
@@ -72,6 +76,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        Gate::authorize('app.roles.edit');
         $modules = Module::all();
         return view('backend.roles.form',compact('role','modules'));
     }
@@ -85,6 +90,7 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        Gate::authorize('app.roles.edit');
         $role->update([
             'name' => $request->name,
             'slug' => Str::slug($request->name),
@@ -102,6 +108,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        Gate::authorize('app.roles.destroy');
         if ($role->deletable) {
             $role->delete();
             notify()->success("Role Successfully Deleted", "Deleted");
